+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Add `show` option to `WeekNumber` config used in `CalenderMonth`
 * Add new highlight states to `DayStateHighlight`: `selectedStart` and `selectedEnd`. For compatibility reasons the selected days hove both the old `selected` AND `selectedStart` and/or `selectedEnd`
 * Add `borderColor` (and `borderColorDisabled`) to `ButtonTheme`
+* Add `cellSpacingPx` to `CalendarMonthTheme`
+* Add `HeaderLeftIcon` and `HeaderRightIcon` to `CalendarMonthTheme` to be able to use custom icons in month switcher buttons
+* Add `borderColor`, `borderRadius` and `rangeBorderRadius` to `CalendarDay`'s `wrapperStyle`
+* Add `rangeBackground`, `verticalExpand` and `horizontalExpand` to `CalendarDay`'s `innerWrapperStyle`
+* Add `rangeTextColor` to `CalendarDay`'s `textProps`
 
 #### 0.0.34